requirements.txt
django>=1.7,<1.8
Pillow>=2.2
django-simple-captcha
django-blog-zinnia
django-mptt
django-tagging
beautifulsoup4
pytz
South
django-xmlrpc
psycopg2
django-suit
zinnia-theme-foundation
Markdown>=2.3.1
easy_thumbnails
django-filer
dealer
django-debug-toolbar
django-contrib-comments>=1.5
raven
coverage
flake8
